feat(node): remove parsing
