+++ crates/jstz_node/Cargo.toml
@@ -33,6 +33,7 @@ jstz_core = { path = "../jstz_core" }
 jstz_crypto = { path = "../jstz_crypto" }
 jstz_proto = { path = "../jstz_proto", features = ["kernel"] }
 jstz_utils = { path = "../jstz_utils" }
+jstz_kernel = { path = "../jstz_kernel" }
 
 log.workspace = true
 mockito.workspace = true
