+++ crates/jstz_node/src/sequencer/queue.rs
@@ -1,6 +1,6 @@
 use std::collections::VecDeque;
 
-use super::inbox::parsing::ParsedInboxMessage;
+use jstz_kernel::inbox::ParsedInboxMessage;
 
 pub struct OperationQueue {
     capacity: usize,
