+++ crates/jstz_node/src/services/operations.rs
@@ -3,11 +3,11 @@ use std::path;
 use std::sync::Arc;
 use std::sync::RwLock;
 
-use crate::sequencer::inbox::parsing::Message;
-use crate::sequencer::inbox::parsing::ParsedInboxMessage;
 use crate::sequencer::queue::OperationQueue;
 use crate::services::accounts::get_account_nonce;
 use crate::RunMode;
+use jstz_kernel::inbox::Message;
+use jstz_kernel::inbox::ParsedInboxMessage;
 
 use super::error::{ServiceError, ServiceResult};
 use super::utils::StoreWrapper;
@@ -293,7 +293,6 @@ mod tests {
     use tezos_crypto_rs::hash::ContractKt1Hash;
     use tower::ServiceExt;
 
-    use crate::sequencer::inbox::parsing::{Message, ParsedInboxMessage};
     use crate::services::utils::StoreWrapper;
     use crate::{
         services::{
@@ -304,6 +303,8 @@ mod tests {
         utils::tests::{dummy_receipt, mock_app_state},
         RunMode,
     };
+    use jstz_kernel::inbox::Message;
+    use jstz_kernel::inbox::ParsedInboxMessage;
 
     use super::MAX_DIRECT_OPERATION_SIZE;
